Fifth Commit
Brain is dead
Removed the grid component of the building system and purely have bricks being built whenever inside the correct outline. Probably a better way to do it, but I don't know that my understanding of the Unity Grid component is great enough for me to do anything different then this solution.