Thirteenth Commit
Fixes issue where 2x2 and 2x1 bricks would clip into the ground when getting placed.
Fixes issue where 2x1 bricks would not be following the correct rotation based on the outline that they are being placed on.
General fixes to improve responsiveness to the player system. Added conditions that made it a lot harder to encounter weird situations.