Beta Progress: Commit Four
Made bricks feel better, changed the gravity to be higher. Fixed an issue when the player would pick up a brick that was being held by Tiny. The brick would be stuttering and have an irregular movement pattern. Instead, the brick goes to the specified spot and the brick's gravity is disabled.